Impl regras de negocio

+++ sql/init/ddl.sql
@@ -1,14 +1,14 @@
 CREATE TABLE users (
-  id SERIAL PRIMARY KEY,
-  balance INTEGER NOT NULL,
-  balance_limit INTEGER NOT NULL
+  id            SERIAL PRIMARY KEY,
+  balance       BIGINT NOT NULL,
+  balance_limit BIGINT NOT NULL
 );
 
 CREATE TABLE transactions (
-    id SERIAL PRIMARY KEY,
-    user_id INTEGER NOT NULL,
-    amount INTEGER NOT NULL,
+    id          SERIAL PRIMARY KEY,
+    user_id     INTEGER NOT NULL,
+    amount      BIGINT NOT NULL,
     description VARCHAR(10) NOT NULL,
-    created_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
-    ttype CHAR(1) NOT NULL
+    created_at  TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
+    ttype       CHAR(1) NOT NULL
 );
